refactor: move builtin app sources
