+++ apps/bin-src/bash.ts
@@ -1,0 +1,164 @@
+export const BASH_SOURCE = `
+  async (syscall, argv) => {
+    const STDOUT_FD = 1;
+    const STDERR_FD = 2;
+    const encode = (s) => new TextEncoder().encode(s);
+    const decode = (b) => new TextDecoder().decode(b);
+
+    async function readFile(path) {
+      const fd = await syscall('open', path, 'r');
+      let out = '';
+      while (true) {
+        const chunk = await syscall('read', fd, 1024);
+        if (chunk.length === 0) break;
+        out += decode(chunk);
+      }
+      await syscall('close', fd);
+      return out;
+    }
+
+    async function readLine(fd) {
+      let line = '';
+      while (true) {
+        const chunk = await syscall('read', fd, 1);
+        if (chunk.length === 0) break;
+        const ch = decode(chunk);
+        if (ch === '\n') break;
+        line += ch;
+      }
+      return line;
+    }
+
+    async function waitPid(pid) {
+      while (true) {
+        const list = await syscall('ps');
+        const proc = list.find(p => p.pid === pid);
+        if (!proc || proc.exited) break;
+        await new Promise(r => setTimeout(r, 50));
+      }
+    }
+
+    const ttyName = argv[0] || 'tty0';
+    let tty;
+    try {
+      tty = await syscall('open', '/dev/' + ttyName, 'r');
+    } catch {
+      await syscall('write', STDERR_FD, encode('bash: unable to open tty\n'));
+      return 1;
+    }
+
+    const initLimits = await syscall('set_quota');
+    let quotaMs = initLimits.quotaMs;
+    let quotaMem = initLimits.quotaMem;
+
+    let nextJob = 1;
+    const jobs = [];
+
+    while (true) {
+      await syscall('write', STDOUT_FD, encode('$ '));
+      const line = (await readLine(tty)).trim();
+      if (!line) continue;
+      if (line === 'exit') break;
+
+      if (line === 'jobs') {
+        let list;
+        try {
+          list = await syscall('jobs');
+        } catch {
+          list = jobs;
+        }
+        for (const j of list) {
+          await syscall('write', STDOUT_FD,
+            encode('[' + j.id + '] ' + (j.status || j.state) + ' ' + j.command + '\n'));
+        }
+        continue;
+      }
+
+      if (line.startsWith('fg ')) {
+        const id = parseInt(line.slice(3).trim(), 10);
+        const job = jobs.find(j => j.id === id);
+        if (job) {
+          for (const pid of job.pids) {
+            await waitPid(pid);
+          }
+          job.state = 'Done';
+        }
+        continue;
+      }
+
+      if (line.startsWith('bg ')) {
+        const id = parseInt(line.slice(3).trim(), 10);
+        const job = jobs.find(j => j.id === id);
+        if (job) job.state = 'Running';
+        continue;
+      }
+
+      if (line.startsWith('ulimit')) {
+        const parts = line.split(/\s+/).slice(1);
+        if (parts.length === 0) {
+          await syscall('write', STDOUT_FD, encode('cpu ' + quotaMs + ' mem ' + quotaMem + '\n'));
+        } else {
+          for (let i = 0; i < parts.length; i++) {
+            if (parts[i] === '-t' && parts[i + 1]) {
+              quotaMs = parseInt(parts[i + 1], 10);
+              i++;
+            } else if (parts[i] === '-m' && parts[i + 1]) {
+              quotaMem = parseInt(parts[i + 1], 10);
+              i++;
+            }
+          }
+          await syscall('set_quota', quotaMs, quotaMem);
+        }
+        continue;
+      }
+
+      if (line.startsWith('kill')) {
+        const args = line.slice(4).trim().split(/\s+/).filter(a => a);
+        for (const arg of args) {
+          if (arg.startsWith('%')) {
+            const id = parseInt(arg.slice(1), 10);
+            let list;
+            try {
+              list = await syscall('jobs');
+            } catch {
+              list = jobs;
+            }
+            const job = list.find(j => j.id === id);
+            if (job) {
+              for (const pid of job.pids) {
+                await syscall('kill', pid);
+              }
+            }
+          } else {
+            const pid = parseInt(arg, 10);
+            if (!isNaN(pid)) {
+              await syscall('kill', pid);
+            }
+          }
+        }
+        continue;
+      }
+
+      const bg = line.endsWith('&');
+      const cmd = bg ? line.slice(0, -1).trim() : line;
+      const [name, ...args] = cmd.split(' ');
+      try {
+        const code = await readFile('/bin/' + name);
+        let m;
+        try { m = JSON.parse(await readFile('/bin/' + name + '.manifest.json')); } catch {}
+        const pid = await syscall('spawn', code, { argv: args, syscalls: m ? m.syscalls : undefined, tty: ttyName, quotaMs, quotaMem });
+        const job = { id: nextJob++, pids: [pid], command: cmd, state: 'Running' };
+        jobs.push(job);
+        if (!bg) {
+          await waitPid(pid);
+          job.state = 'Done';
+        }
+      } catch {
+        await syscall('write', STDERR_FD, encode('bash: ' + name + ': command not found\n'));
+      }
+    }
+
+    await syscall('close', tty);
+    return 0;
+  }
+`;
