+++ apps/bin-src/reboot.ts
@@ -1,0 +1,6 @@
+export const REBOOT_SOURCE = `
+  async (syscall) => {
+    await syscall('reboot');
+    return 0;
+  }
+`;
